requirements/base.txt
pytz==2018.3  # https://github.com/stub42/pytz
awesome-slugify==1.6.5  # https://github.com/dimka665/awesome-slugify
Pillow==5.0.0  # https://github.com/python-pillow/Pillow
argon2-cffi==18.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==3.3.1  # https://github.com/evansd/whitenoise
redis>=2.10.5  # https://github.com/antirez/redis

# Django
# ------------------------------------------------------------------------------
django==1.11.8  # pyup: < 2.1  # https://www.djangoproject.com/
django-environ==0.4.4  # https://github.com/joke2k/django-environ
django-model-utils==3.1.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.35.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.9.0  # https://github.com/niwinz/django-redis

# Django REST Framework
djangorestframework==3.7.7  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client

# Your custom requirements go here
channels==2.0.2
django-imagekit==4.0.2
django-widget-tweaks==1.4.1
jsonfield==2.0.2
